requirements.txt
requests
python-telegram-bot==13.5
flask
deta